+++ src/index.ts
@@ -1,17 +1,7 @@
-import {
-  Options,
-  createXMLEditor,
-  EditorFunc,
-  Element,
-  newElement,
-  Selector,
-} from './xml-stream-editor.js'
-
+export { Element, newElement } from './element.js'
 export {
   Options,
   createXMLEditor,
   EditorFunc,
-  Element,
-  newElement,
   Selector,
-}
+} from './xml-stream-editor.js'