+++ src/markup.ts
@@ -1,7 +1,4 @@
 import xmlescape from 'xml-escape'
-import xnv from 'xml-name-validator'
-
-export const isValidName = xnv.qname
 
 export const toAttrValue = (value: string): string => {
   return xmlescape(value)