+++ src/xml-stream-editor.ts
@@ -3,12 +3,13 @@ import { Transform, TransformCallback } from 'node:stream'
 
 import { SaxesOptions, SaxesParser, TagForOptions, XMLDecl } from 'saxes'
 
-import { isValidName, toAttrValue, toBodyText, toCloseTag,
-  toOpenTag } from './markup.js'
+import { Element, ParsedElement } from './element.js'
+import { toAttrValue, toBodyText, toCloseTag, toOpenTag } from './markup.js'
+import { ElementPath, SelectorRule } from './selector.js'
 
 export type Selector = string
 export type EditorFunc = (elm: Element) => Element | undefined
-export type EditingRules = Record<Selector, EditorFunc>
+export type EditingRules = Map<Selector, EditorFunc>
 
 export interface Options {
   // Whether to check and enforce the validity of created and modified
@@ -30,88 +31,14 @@ export interface Options {
 // selector, and the corresponding function the user provided to
 // edit the matching element
 interface ElementToEditInfo {
-  selector: Selector
+  element: ParsedElement
   func: EditorFunc
-  element: Element
+  selector: SelectorRule
 }
 
 interface ParserStackEntry {
-  element: Element
-  path: string
-}
-
-export interface Element {
-  name: string
-  text?: string
-  attributes: Record<string, string>
-  children: Element[]
-}
-
-export const newElement = (name: string): Element => {
-  return {
-    name: name,
-    text: undefined,
-    attributes: Object.create(null),
-    children: [],
-  }
-}
-
-const throwOnInvalidElement = (elm: Element): void => {
-  if (typeof elm.name !== 'string') {
-    throw new Error('No name provided for element')
-  }
-
-  if (!isValidName(elm.name)) {
-    throw new Error(`"${elm.name}" is not a valid XML element name`)
-  }
-
-  if (typeof elm.attributes !== 'object' || elm.attributes === null) {
-    throw new Error('"attributes" property on element is not an object')
-  }
-
-  for (const attrName of Object.keys(elm.attributes)) {
-    if (!isValidName(attrName)) {
-      throw new Error(`"${attrName}" is not a valid XML attribute name`)
-    }
-  }
-
-  elm.children.forEach(throwOnInvalidElement)
-}
-
-const cloneElement = (elm: Element): Element => {
-  const newElm = newElement(elm.name)
-  newElm.text = elm.text
-  newElm.attributes = JSON.parse(JSON.stringify(elm.attributes))
-  newElm.children = elm.children.map(cloneElement)
-  return newElm
-}
-
-const elementForNode = (node: TagForOptions<SaxesOptions>): Element => {
-  // Here we check if each attribute name is simple (and so just a
-  // string), or in the namespace representation the "saxes" library
-  // uses (in which case attrValue will be a SaxesAttributeNS
-  // object, that we have to unpack a bit)
-  const attributes: Record<string, string> = Object.create(null)
-  if (node.attributes) {
-    for (const [attrName, attrValue] of Object.entries(node.attributes)) {
-      if (typeof attrValue === 'string') {
-        attributes[attrName] = attrValue
-        continue
-      }
-      attributes[attrValue.name] = attrValue.value
-    }
-  }
-
-  return elementForNameAndAttrs(node.name, attributes)
-}
-
-const elementForNameAndAttrs = (name: string,
-                                attrs?: Record<string, string>) => {
-  const newElm = newElement(name)
-  if (attrs) {
-    newElm.attributes = attrs
-  }
-  return newElm
+  element: ParsedElement
+  path: ElementPath
 }
 
 class XMLStreamEditorTransformer extends Transform {
@@ -130,9 +57,9 @@ class XMLStreamEditorTransformer extends Transform {
   // check newly parsed elements against the passed editor rules.
   readonly #parseStack: ParserStackEntry[] = []
 
-  // This is a map of (VERY) simple xpaths (i.e., only XML element names;
-  // no attributes, no name spaces, etc).
-  readonly #editingRules: EditingRules
+  // This is a map of objects that represent simple xpaths (i.e., only XML
+  // element names (no attributes, no name spaces, etc).
+  readonly #rules: Map<SelectorRule, EditorFunc>
 
   // Handle to the 'saxes' xml parser object.
   readonly #xmlParser: SaxesParser
@@ -146,11 +73,15 @@ class XMLStreamEditorTransformer extends Transform {
   // can pass it along in the transformer callback next time we get data.
   #error?: Error
 
-  #pushParsedElementToStack (element: Element): void {
+  #pushParsedElementToStack (element: ParsedElement): void {
     const topOfStackElm = this.#parseStack.at(-1)
+    // We prefix every element name in the parse stack with '@' (a character
+    // that isn't valid in an XML element name) so that we can easily
+    // check if a selector matches the parse stack by just checking if
+    // the selector matches right end of the stack path.
     const pathToElement = topOfStackElm
-      ? topOfStackElm.path + ' ' + element.name
-      : element.name
+      ? topOfStackElm.path.append(element.name)
+      : new ElementPath(element.name)
     this.#parseStack.push({
       element: element,
       path: pathToElement,
@@ -165,15 +96,14 @@ class XMLStreamEditorTransformer extends Transform {
     // This method is only called after pushing an element to the stack,
     // so this is guaranteed to be true
     assert(topOfStack)
-    const topOfStackPath = topOfStack.path
 
-    for (const [selector, editorFunc] of Object.entries(this.#editingRules)) {
-      if (topOfStackPath.endsWith(selector)) {
+    for (const [selectorRule, editorFunc] of this.#rules.entries()) {
+      if (topOfStack.path.matches(selectorRule)) {
         // The depth of the root of this subtree in the stack
         const depth = this.#parseStack.length - 1
         assert(depth >= 0)
         const elmToEdit = this.#parseStack[depth].element
-        return { selector: selector, func: editorFunc, element: elmToEdit }
+        return { selector: selectorRule, func: editorFunc, element: elmToEdit }
       }
     }
     return null
@@ -196,13 +126,17 @@ class XMLStreamEditorTransformer extends Transform {
 
   #callUserFuncOnCompletedElementAndWriteToStream (): void {
     assert(this.#elmToEditInfo)
-    const clonedElm = cloneElement(this.#elmToEditInfo.element)
+    const clonedElm = this.#elmToEditInfo.element.clone()
+    const editElmFunc = this.#elmToEditInfo.func
 
     try {
-      const editedElm = this.#elmToEditInfo.func(clonedElm)
+      const editedElm = editElmFunc(clonedElm)
       if (editedElm) {
         if (this.#options.validate === true) {
-          throwOnInvalidElement(editedElm)
+          const [isValid, error] = editedElm.validate()
+          if (!isValid) {
+            throw error
+          }
         }
         this.#writeElementToStream(editedElm)
       }
@@ -225,7 +159,7 @@ class XMLStreamEditorTransformer extends Transform {
       //    and append ourselves to the stack.
       // 3. We are NOT the root of a subtree to be edited, in which case
       //    we just add ourselves to the stack.
-      const newElement = elementForNode(node)
+      const newElement = ParsedElement.fromSaxesNode(node)
       this.#pushParsedElementToStack(newElement)
       // Check for case one
       if (this.#isInSubtreeToBeEdited()) {
@@ -327,7 +261,13 @@ class XMLStreamEditorTransformer extends Transform {
     }
 
     this.#options = mergedOptions
-    this.#editingRules = editingRules
+    this.#rules = new Map<SelectorRule, EditorFunc>()
+    for (const [selector, editFunc] of Object.entries(editingRules)) {
+      // This will throw if one of the user-provided selectors
+      // is invalid.
+      const parsedSelector = new SelectorRule(selector)
+      this.#rules.set(parsedSelector, editFunc)
+    }
     this.#xmlParser = new SaxesParser(this.#options.saxes)
     this.#configureParserCallbacks()
   }